pyproject.toml
[project]
name = "coordinated"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Arvin Kushwaha", email = "arvin.singh.kushwaha@gmail.com" }
]
dependencies = [
    "jax>=0.4.30",
    "jaxlib[cuda12_pip] @ https://storage.googleapis.com/jax-releases/cuda12/jaxlib-0.4.28+cuda12.cudnn89-cp312-cp312-manylinux2014_x86_64.whl",
    "optax>=0.2.3",
    "flashbax>=0.1.2",
    "equinox>=0.11.4",
    "jaximal @ git+https://github.com/ArvinSKushwaha/jaximal.git",
    "matplotlib-backend-sixel @ git+https://github.com/ArvinSKushwaha/matplotlib-backend-sixel",
]
readme = "README.md"
requires-python = ">= 3.12"
license = { text = "MIT" }

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "matplotlib>=3.9.1",
    "pyqt5>=5.15.11",
    "pycairo>=1.26.1",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint.isort]
lines-between-types = 1
relative-imports-order = "closest-to-furthest"

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.basedpyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownParameterType = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownLambdaType = false
reportUntypedFunctionDecorator = false
